Toast string change
Currently when you attempt to configure anything while the equalizer is turned off,
you see a toast appear letting you know to press the so called 'power button' on the
top right. This shit doesn't sound like it was written by a native english speaker.

The switch to enable the equalizer is not a 'power button', is a switch. Saying power
button is going to confuse the living hell out of people.

Change-Id: I51098b03a1e93efe783e33e43e2ab59af9c52261